Update latex.py filter empty figures
Some figures are examples from the template that people do not always clean up. The code removes the filenames when the source is not found but keeps the rest. Sometimes, these have the same label, which leads to having them selected.

The patch removes the figures without a valid source image